+++ data-raw/create_constants.R
@@ -37,7 +37,8 @@ usethis::use_data(psut_cols, overwrite = TRUE)
 
 aggregate_cols <- list(aggregate_primary = "EX.p",
                        net_aggregate_demand = "EX.d_net",
-                       gross_aggregate_demand = "EX.d_gross")
+                       gross_aggregate_demand = "EX.d_gross",
+                       region = "Region")
 usethis::use_data(aggregate_cols, overwrite = TRUE)